use narrow_cast when casting up to size_t

+++ ThrEd4/form.cpp
@@ -845,7 +845,7 @@ void form::setmfrm() {
 	const auto offset    = POINT{ Msg.pt.x - StitchWindowOrigin.x - point.x + dToL(FormMoveDelta.x),
                                Msg.pt.y - StitchWindowOrigin.y - point.y + dToL(FormMoveDelta.y) };
 	auto&      formLines = *FormLines;
-	formLines.resize(gsl::narrow_cast<size_t>(closeForm.vertexCount) + 1);
+	formLines.resize(gsl::narrow_cast<size_t>(closeForm.vertexCount) + 1u);
 	for (auto iForm = 0u; iForm < closeForm.vertexCount; iForm++) {
 		form::sfCor2px(vertexIt[iForm], point);
 		formLines[iForm].x = point.x + offset.x;
@@ -2250,7 +2250,7 @@ void form::internal::fnvrt(std::vector<fPOINT>*   currentFillVertices,
 	const auto step          = (highX - lowX) / fillLineCount;
 	auto       currentX      = lowX;
 	auto       projectedPoints = std::vector<dPOINTLINE>{};
-	projectedPoints.reserve(gsl::narrow_cast<size_t>(VertexCount) + 2);
+	projectedPoints.reserve(gsl::narrow_cast<size_t>(VertexCount) + 2u);
 	for (auto iLine = 0u; iLine < fillLineCount; iLine++) {
 		auto iLineCounter = 0u;
 		currentX += step;
@@ -2263,7 +2263,7 @@ void form::internal::fnvrt(std::vector<fPOINT>*   currentFillVertices,
 		}
 		fillLineCount += iLineCounter;
 	}
-	lineEndpoints.reserve(gsl::narrow_cast<size_t>(fillLineCount) + 1);
+	lineEndpoints.reserve(gsl::narrow_cast<size_t>(fillLineCount) + 1u);
 	auto lineGroupIndex = 0u;
 	// groupIndex cannot be more than fillLineCount so reserve that amount of memory to reduce re-allocations
 	groupIndexSequence.reserve(fillLineCount);
@@ -2414,9 +2414,9 @@ void form::internal::plbak(unsigned int backPoint) noexcept {
 void form::internal::plbrd(double edgeSpacing, FRMHED& angledForm) {
 	// Ensure that we have at least 4 array members
 	auto fillVerticalRect = std::vector<VRCT2>{};
-	fillVerticalRect.resize(gsl::narrow_cast<size_t>(VertexCount) + 3);
+	fillVerticalRect.resize(gsl::narrow_cast<size_t>(VertexCount) + 3u);
 	auto underlayVerticalRect = std::vector<VRCT2>{};
-	underlayVerticalRect.resize(gsl::narrow_cast<size_t>(VertexCount) + 3);
+	underlayVerticalRect.resize(gsl::narrow_cast<size_t>(VertexCount) + 3u);
 	prebrd(angledForm);
 	satin::satout(SelectedForm->borderSize);
 	InsidePoints->push_back((*InsidePoints)[0]);
@@ -2904,7 +2904,7 @@ unsigned form::internal::insect(std::vector<CLIPSORT>&     clipIntersectData,
 		auto iDestination = 1u;
 		for (iIntersection = 0; iIntersection < count - 1; iIntersection++) {
 			if (fabs(gsl::narrow_cast<double>(arrayOfClipIntersectData[iIntersection]->segmentLength)
-			         - arrayOfClipIntersectData[gsl::narrow_cast<size_t>(iIntersection) + 1]->segmentLength)
+			         - arrayOfClipIntersectData[gsl::narrow_cast<size_t>(iIntersection) + 1u]->segmentLength)
 			    > TINY) {
 				mvpclp(arrayOfClipIntersectData, iDestination++, iIntersection + 1);
 			}
@@ -3208,7 +3208,7 @@ void form::internal::clpcon(const std::vector<RNGCNT>& textureSegments, std::vec
 		bool breakFlag = false;
 		for (auto iRegion = 0u; iRegion < iclpxSize; iRegion++) {
 			auto regionCrossingStart = iclpx[iRegion];
-			auto regionCrossingEnd   = iclpx[gsl::narrow_cast<size_t>(iRegion) + 1];
+			auto regionCrossingEnd   = iclpx[gsl::narrow_cast<size_t>(iRegion) + 1u];
 			pasteLocation.x          = clipWidth * (iRegion + clipGrid.left);
 			auto  clipVerticalOffset = 0.0f;
 			auto  lineSegmentStart   = fPOINT{}; // vertical clipboard line segment start
@@ -3590,14 +3590,14 @@ bool form::internal::lnclos(std::vector<unsigned>& groupIndexSequence,
 		return false;
 	}
 	if (lineEndPoint0 != nullptr) {
-		auto count0 = (groupIndexSequence[gsl::narrow_cast<size_t>(group0) + 1] - groupIndexSequence[group0]) >> 1;
+		auto count0 = (groupIndexSequence[gsl::narrow_cast<size_t>(group0) + 1u] - groupIndexSequence[group0]) >> 1;
 		auto index0 = 0u;
 		while ((count0 != 0u) && lineEndPoint0[index0].line != line0) {
 			count0--;
 			index0 += 2;
 		}
 		if (count0 != 0u) {
-			auto count1 = (groupIndexSequence[gsl::narrow_cast<size_t>(group1) + 1] - groupIndexSequence[group1]) >> 1;
+			auto count1 = (groupIndexSequence[gsl::narrow_cast<size_t>(group1) + 1u] - groupIndexSequence[group1]) >> 1;
 			auto index1 = 0u;
 			if (const auto lineEndPoint1 = &lineEndpoints[groupIndexSequence[group1]]) {
 				while ((count1 != 0u) && lineEndPoint1[index1].line != line1) {
@@ -3722,11 +3722,11 @@ bool form::internal::notdun(std::vector<RGSEQ>&            tempPath,
 
 	const auto regionPath = &tempPath[sequencePathIndex];
 	regionPath[0].pcon    = mapIndexSequence[doneRegion];
-	regionPath[0].count   = mapIndexSequence[gsl::narrow_cast<size_t>(doneRegion) + 1] - regionPath[0].pcon;
+	regionPath[0].count   = mapIndexSequence[gsl::narrow_cast<size_t>(doneRegion) + 1u] - regionPath[0].pcon;
 	for (auto iPath = 1u; iPath < level; iPath++) {
 		regionPath[iPath].pcon = mapIndexSequence[pathMap[regionPath[iPath - 1].pcon].node];
 		regionPath[iPath].count
-		    = mapIndexSequence[gsl::narrow_cast<size_t>(pathMap[regionPath[gsl::narrow_cast<size_t>(iPath) - 1].pcon].node) + 1]
+		    = mapIndexSequence[gsl::narrow_cast<size_t>(pathMap[regionPath[gsl::narrow_cast<size_t>(iPath) - 1u].pcon].node) + 1u]
 		      - regionPath[iPath].pcon;
 	}
 	while (visitedRegions[pathMap[regionPath[previousLevel].pcon].node] && previousLevel >= 0) {
@@ -3751,7 +3751,7 @@ bool form::internal::notdun(std::vector<RGSEQ>&            tempPath,
 					regionPath[pivot].pcon = mapIndexSequence[pathMap[regionPath[pivot - 1].pcon].node];
 					regionPath[pivot].count
 					    = mapIndexSequence
-					          [gsl::narrow_cast<size_t>(pathMap[regionPath[gsl::narrow_cast<size_t>(pivot) - 1].pcon].node) + 1]
+					          [gsl::narrow_cast<size_t>(pathMap[regionPath[gsl::narrow_cast<size_t>(pivot) - 1u].pcon].node) + 1u]
 					      - regionPath[pivot].pcon;
 				}
 				else {
@@ -3864,9 +3864,9 @@ void form::internal::nxtseq(std::vector<FSEQ>&           sequencePath,
                             unsigned&                    pathCount) {
 	auto iPath = mapIndexSequence[sequencePath[pathIndex].node];
 
-	if ((gsl::narrow_cast<size_t>(pathIndex) + 1) < sequencePath.size()) {
-		const auto nextNode = sequencePath[gsl::narrow_cast<size_t>(pathIndex) + 1].node;
-		while (iPath < mapIndexSequence[gsl::narrow_cast<size_t>(sequencePath[pathIndex].node) + 1]
+	if ((gsl::narrow_cast<size_t>(pathIndex) + 1u) < sequencePath.size()) {
+		const auto nextNode = sequencePath[gsl::narrow_cast<size_t>(pathIndex) + 1u].node;
+		while (iPath < mapIndexSequence[gsl::narrow_cast<size_t>(sequencePath[pathIndex].node) + 1u]
 		       && pathMap[iPath].node != nextNode) {
 			iPath++;
 		}
@@ -4030,7 +4030,7 @@ void form::internal::duseq(const std::vector<SMALPNTL*>& sortedLines,
 					else {
 						if (savedTopLine != sortedLines[iLineDec][1].line) {
 							if (iLineDec != 0u) {
-								sequenceLines = duseq2(sortedLines[gsl::narrow_cast<size_t>(iLineDec) + 1]);
+								sequenceLines = duseq2(sortedLines[gsl::narrow_cast<size_t>(iLineDec) + 1u]);
 							}
 							flag          = true;
 							sequenceLines = duseq2(sortedLines[iLineDec]);
@@ -4040,7 +4040,7 @@ void form::internal::duseq(const std::vector<SMALPNTL*>& sortedLines,
 				}
 				else {
 					if (StateMap.testAndReset(StateFlag::SEQDUN)) {
-						sequenceLines = duseq2(sortedLines[gsl::narrow_cast<size_t>(iLineDec) + 1]);
+						sequenceLines = duseq2(sortedLines[gsl::narrow_cast<size_t>(iLineDec) + 1u]);
 					}
 					flag          = true;
 					sequenceLines = sortedLines[iLineDec];
@@ -4197,7 +4197,7 @@ void form::internal::durgn(const std::vector<FSEQ>&      sequencePath,
 		seqn = sequenceEnd;
 	}
 	if (sortedLines[seql]->group != lastGroup) {
-		if (seql < sequenceEnd && sortedLines[gsl::narrow_cast<size_t>(seql) + 1]->group == lastGroup) {
+		if (seql < sequenceEnd && sortedLines[gsl::narrow_cast<size_t>(seql) + 1u]->group == lastGroup) {
 			seql++;
 		}
 		else {
@@ -4218,7 +4218,7 @@ void form::internal::durgn(const std::vector<FSEQ>&      sequencePath,
 		}
 	}
 	if (sortedLines[seqn]->group != nextGroup) {
-		if (seqn < sequenceEnd && sortedLines[gsl::narrow_cast<size_t>(seqn) + 1]->group == nextGroup) {
+		if (seqn < sequenceEnd && sortedLines[gsl::narrow_cast<size_t>(seqn) + 1u]->group == nextGroup) {
 			seqn++;
 		}
 		else {
@@ -4383,14 +4383,14 @@ void form::internal::lcon(std::vector<unsigned>& groupIndexSequence, std::vector
 		BSequence->clear();
 
 		auto mapIndexSequence = std::vector<unsigned>{};
-		mapIndexSequence.reserve(gsl::narrow_cast<size_t>(regionCount) + 1);
+		mapIndexSequence.reserve(gsl::narrow_cast<size_t>(regionCount) + 1u);
 		auto pathMap      = std::vector<RCON>{};
 		auto sequencePath = std::vector<FSEQ>{};
 
 		if (regionCount > 1) {
 			auto pathMapIndex = 0u;
 			// use the number of possible pairs of nodes n(n - 1)/2 and account for RegionCount possibly being odd
-			pathMap.reserve(gsl::narrow_cast<size_t>((regionCount * (regionCount - 1)) / 2) + 2);
+			pathMap.reserve(gsl::narrow_cast<size_t>((regionCount * (regionCount - 1u)) / 2u) + 2u);
 			for (auto iSequence = 0u; iSequence < regionCount; iSequence++) {
 				mapIndexSequence.push_back(pathMapIndex);
 				auto count              = 0;
@@ -4449,7 +4449,7 @@ void form::internal::lcon(std::vector<unsigned>& groupIndexSequence, std::vector
 			}
 			OutputIndex   = 0;
 			auto tempPath = std::vector<RGSEQ>{};
-			tempPath.resize(gsl::narrow_cast<size_t>((regionCount * (regionCount - 1)) / 2) + 1);
+			tempPath.resize(gsl::narrow_cast<size_t>((regionCount * (regionCount - 1u)) / 2u) + 1u);
 
 			// find the leftmost region in pathMap
 			auto sequencePathIndex = 1u;
@@ -4765,9 +4765,9 @@ void form::internal::clpfm() {
 	ActivePointIndex = 0;
 	for (auto iSequence = 0u; iSequence < SequenceIndex - 2; iSequence += 2) {
 		auto& bSeq0 = (*BSequence)[iSequence];
-		auto& bSeq1 = (*BSequence)[gsl::narrow_cast<size_t>(iSequence) + 1];
-		auto& bSeq2 = (*BSequence)[gsl::narrow_cast<size_t>(iSequence) + 2];
-		auto& bSeq3 = (*BSequence)[gsl::narrow_cast<size_t>(iSequence) + 3];
+		auto& bSeq1 = (*BSequence)[gsl::narrow_cast<size_t>(iSequence) + 1u];
+		auto& bSeq2 = (*BSequence)[gsl::narrow_cast<size_t>(iSequence) + 2u];
+		auto& bSeq3 = (*BSequence)[gsl::narrow_cast<size_t>(iSequence) + 3u];
 		const auto leftLength  = hypot(bSeq1.x - bSeq0.x, bSeq1.y - bSeq0.y);
 		const auto rightLength = hypot(bSeq3.x - bSeq2.x, bSeq3.y - bSeq2.y);
 		const auto leftDelta   = dPOINT{ bSeq1.x - bSeq0.x, bSeq1.y - bSeq0.y };
@@ -5288,7 +5288,7 @@ bool form::chkfrm(std::vector<POINT>& stretchBoxLine, double& xyRatio) {
 		if (minimumLength < CLOSENUF) {
 			form::ritfrct(ClosestFormToCursor, StitchWindowDC);
 			for (auto iCorner = 0u; iCorner < 4; iCorner++) {
-				stretchBoxLine[iCorner] = formControls[gsl::narrow<size_t>(iCorner) << 1];
+				stretchBoxLine[iCorner] = formControls[gsl::narrow_cast<size_t>(iCorner) << 1];
 			}
 			stretchBoxLine[4] = stretchBoxLine[0];
 			thred::strtchbox(stretchBoxLine);
@@ -7148,7 +7148,7 @@ void form::internal::doTimeWindow(float rangeX, const std::vector<unsigned>& xPo
 	for (auto iColumn = 1u; iColumn < rangeX - checkLength - 1; iColumn++) {
 		snpfn(xPoints,
 		      xHistogram[iColumn],
-		      xHistogram[gsl::narrow_cast<size_t>(iColumn) + 1],
+		      xHistogram[gsl::narrow_cast<size_t>(iColumn) + 1u],
 		      xHistogram[gsl::narrow_cast<size_t>(iColumn) + checkLength]);
 		Polyline(timeDC, formLines.data(), 2);
 		timePosition += timeStep;
@@ -7164,7 +7164,7 @@ void form::internal::snp(unsigned start, unsigned finish) {
 	auto xPoints = std::vector<unsigned>{};
 	xPoints.resize(PCSHeader.stitchCount);
 	auto xHistogram = std::vector<unsigned>{};
-	xHistogram.resize(gsl::narrow<size_t>(std::round(range.x)) + 1);
+	xHistogram.resize(gsl::narrow<size_t>(std::round(range.x)) + 1u);
 
 	const auto attribute = (ClosestFormToCursor << 4) & FRMSK;
 	if (StateMap.test(StateFlag::FORMSEL)) {
@@ -8128,7 +8128,7 @@ constexpr unsigned form::internal::duat(unsigned attribute) {
 void form::internal::srtf(const std::vector<fPOINTATTR>& tempStitchBuffer, unsigned start, unsigned finish) {
 	if (start != finish) {
 		auto stitchHistogram = std::vector<unsigned>{};
-		stitchHistogram.resize(gsl::narrow<size_t>(FormList->size()) << 2);
+		stitchHistogram.resize(FormList->size() << 2);
 		for (auto iStitch = start; iStitch < finish; iStitch++) {
 			stitchHistogram[duat(tempStitchBuffer[iStitch].attribute)]++;
 		}
@@ -8174,7 +8174,7 @@ void form::srtbyfrm() {
 		}
 		fi::srtf(tempStitchBuffer, 0, colorHistogram[0]);
 		for (auto iColor = 0u; iColor < 15; iColor++) {
-			fi::srtf(tempStitchBuffer, colorHistogram[iColor], colorHistogram[gsl::narrow_cast<size_t>(iColor) + 1]);
+			fi::srtf(tempStitchBuffer, colorHistogram[iColor], colorHistogram[gsl::narrow_cast<size_t>(iColor) + 1u]);
 		}
 	}
 	else {
